Comentando y limpiando el codigo

+++ src/funcionesProductos/FuncionesDescuento.java
@@ -4,7 +4,7 @@ public interface FuncionesDescuento {
 
     public void setDescuento(Integer descuento);
 
-    public Integer getDescuento();
+    public Float getDescuento();
 
     public static Float calcularDescuento(Float precio, int descuento) {
         System.out.println("\n\n\nEN LA FUNCION CALCULARdESCUENTO: " + ((precio * descuento)));
